+++ Expanded Cheat Menu/items.lua
@@ -1,0 +1,32 @@
+return {
+	PlaceObj("ModItemOptionToggle", {
+		"name", "HideCheatsMenu",
+		"DisplayName", T(0000, "Hide Cheats Menu"),
+		"Help", T(0000, "Hide a menu entry to reduce horizontal space."),
+		"DefaultValue", false,
+	}),
+	PlaceObj("ModItemOptionToggle", {
+		"name", "HideECMMenu",
+		"DisplayName", T(0000, "Hide ECM Menu"),
+		"Help", T(0000, "Hide a menu entry to reduce horizontal space."),
+		"DefaultValue", false,
+	}),
+	PlaceObj("ModItemOptionToggle", {
+		"name", "HideGameMenu",
+		"DisplayName", T(0000, "Hide Game Menu"),
+		"Help", T(0000, "Hide a menu entry to reduce horizontal space."),
+		"DefaultValue", false,
+	}),
+	PlaceObj("ModItemOptionToggle", {
+		"name", "HideDebugMenu",
+		"DisplayName", T(0000, "Hide Debug Menu"),
+		"Help", T(0000, "Hide a menu entry to reduce horizontal space."),
+		"DefaultValue", false,
+	}),
+	PlaceObj("ModItemOptionToggle", {
+		"name", "HideHelpMenu",
+		"DisplayName", T(0000, "Hide Help Menu"),
+		"Help", T(0000, "Hide a menu entry to reduce horizontal space."),
+		"DefaultValue", false,
+	}),
+}