+++ Mods ChoGGi/Disable Achievements/Code/Script.lua
@@ -1,7 +1,4 @@
 -- See LICENSE for terms
 
--- wtf? fix for this error:
--- Error loading file PackedMods/*****/Code/Script.lua: PackedMods/*****/Code/Script.lua:1: syntax error near '<\1>'
-
--- all this does is replace the func that unlocks them with nadda
+-- replace the func that unlocks them with nadda
 AchievementUnlock = empty_func