+++ Mods ChoGGi/Game Rules Breakthroughs/metadata.lua
@@ -14,6 +14,7 @@ return PlaceObj("ModDef", {
 	},
 	"has_options", true,
 	"description", [[Adds all breakthroughs to game rules (selected rules will be unlocked).
+Sister mod: [url=https://steamcommunity.com/sharedfiles/filedetails/?id=2324443848]Unlock Breakthroughs[/url]
 
 Mod Options:
 Breakthroughs Researched: Research instead of unlock breakthroughs (New game needed after option is applied or maybe restart game).