+++ Mods ChoGGi/Remove Console Dome Cap/Code/Script.lua
@@ -1,7 +1,4 @@
 -- See LICENSE for terms
 
--- wtf? fix for this error:
--- Error loading file PackedMods/*****/Code/Script.lua: PackedMods/*****/Code/Script.lua:1: syntax error near '<\1>'
-
---~ const.DomeCap = max_int
+-- nil is close enough to false
 AreDomesCapped = empty_func