+++ Mods ChoGGi/Stop Death Failure/Code/Script.lua
@@ -1,6 +1,4 @@
 -- See LICENSE for terms
 
--- wtf? fix for this error:
--- Error loading file PackedMods/*****/Code/Script.lua: PackedMods/*****/Code/Script.lua:1: syntax error near '<\1>'
-
+-- replace the func that gameovers with nadda (it's supposed to return false, but nil is good enough)
 IsGameOver = empty_func