+++ Mods ChoGGi/Unlock Breakthroughs/metadata.lua
@@ -1,0 +1,25 @@
+return PlaceObj("ModDef", {
+	"title", "Unlock Breakthroughs",
+	"id", "ChoGGi_UnlockBreakthroughs",
+	"lua_revision", 249143,
+	"steam_id", "2324443848",
+	"pops_any_uuid", "36ec1f1c-283a-419c-9fa6-d7c640a81967",
+	"version", 1,
+	"version_major", 0,
+	"version_minor", 1,
+	"image", "Preview.jpg",
+	"author", "ChoGGi",
+	"code", {
+		"Code/Script.lua",
+	},
+	"has_options", true,
+	"TagGameplay", true,
+	"TagOther", true,
+	"description", [[Unlock breakthroughs in mod options.
+Sister mod: [url=https://steamcommunity.com/sharedfiles/filedetails/?id=1753933193]Game Rules Breakthroughs[/url]
+
+Other Mod Options:
+Breakthroughs Researched: Turn on to research instead of unlock breakthroughs.
+Always Apply Options: Unlock/Research Breakthroughs whenever you load a game/start a new game.
+]],
+})