+++ Mods ChoGGi/Wide Area Forestation/items.lua
@@ -6,7 +6,7 @@ return {
 		"DisplayName", T(302535920011643, "Max Size"),
 		"Help", T(302535920011644, [[Set the max size of the forestation area.
 
-Warning:  Setting this above 200 can cause crashing, save before raising.]]),
+Warning: Setting this above 200 can cause crashing, save before raising.]]),
 		"DefaultValue", 200,
 		"MinValue", 10,
 		"MaxValue", 500,