+++ Expanded Cheat Menu/Changelog.md
@@ -1,6 +1,9 @@
 ## ECM Changelog
 
 ## v15.8 (Unreleased)
+### Added:
+- Cheat menu entries can now be individually hidden in mod options.
+- UnpublishParadoxMod() func: You don't need to use mod editor to remove a mod.
 
 ## v15.7 (08 Dec 2020)
 ### Changed: